fix: move normalize_urls and add_code_block_language_hint to feishu plugin
Move URL normalization and code block language hint from generic
content_normalizer to feishu IM plugin's parse stage, so these
platform-specific transformations run at message ingestion rather
than in the post-processing pipeline.

Source: user
Type: fix
