+++ src/processor_chain/content_normalizer_tests.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::im_adapter::normalized::{add_code_block_language_hint, normalize_urls};
 use crate::processor_chain::context::{MessageContext, RawMessage};
 use crate::processor_chain::processor::MessageProcessor;
